Move 8.18 tests into 8_18_test.go
