+++ functionaltests/8_18_test.go
@@ -79,3 +79,59 @@ func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
 		})
 	}
 }
+
+func TestUpgrade_8_18_to_9_0_Snapshot(t *testing.T) {
+	t.Parallel()
+	from := getLatestSnapshot(t, "8.18")
+	to := getLatestSnapshot(t, "9.0")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := basicUpgradeILMTestScenarios(
+		from.Version,
+		to.Version,
+		apmErrorLogs{
+			tlsHandshakeError,
+			esReturnedUnknown503,
+			refreshCache503,
+			// TODO: remove once fixed
+			populateSourcemapFetcher403,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
+
+func TestUpgrade_8_18_to_9_0_BC(t *testing.T) {
+	t.Parallel()
+	from := getLatestVersionOrSkip(t, "8.18")
+	to := getLatestBCOrSkip(t, "9.0")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := basicUpgradeILMTestScenarios(
+		from.Version,
+		to.Version,
+		apmErrorLogs{
+			tlsHandshakeError,
+			esReturnedUnknown503,
+			refreshCache503,
+			// TODO: remove once fixed
+			populateSourcemapFetcher403,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
